+++ src/error/DiscoveryFailedError.ts
@@ -1,0 +1,5 @@
+export default class DiscoveryFailedError extends Error {
+    constructor(options?: ErrorOptions) {
+        super("Failed to discover server capabilities", options);
+    }
+}