+++ src/schemas/gamerule/UntypedGameRule.ts
@@ -1,5 +1,5 @@
 import GameRule from "./GameRule";
 
-export default class UntypedGameRule extends GameRule<string> {
+export default class UntypedGameRule extends GameRule<string|boolean|number> {
 
 }